refactor: pkgReader, pkgUploader
